agregue lo faltante

+++ condicionales.py
@@ -1,5 +1,4 @@
-# --- Bloque 1: Verificación de mayoría de edad ---
-# Solicitamos un número y lo convertimos a entero (int)
+
 edad_usuario = int(input("Ingresa tu edad: "))
 
 if edad_usuario >= 18:
@@ -8,9 +7,8 @@ else:
     print("Eres menor de edad")
 
 
-# --- Bloque 2: Evaluación de calificaciones con elif ---
-# Usamos float por si la calificación tiene decimales (ej: 5.5)
-nota = float(input("Ingresa tu calificación (1 a 7): "))
+
+nota = int(input("Ingresa tu calificación (1 a 7): "))
 
 if nota == 7:
     print("Excelente")
@@ -24,8 +22,7 @@ else:
     print("Insuficiente")
 
 
-# --- Bloque 3: Condiciones anidadas (Número positivo, cero o negativo) ---
-# Un if dentro de otro para separar primero los no-negativos
+
 numero_entero = int(input("Ingresa un número entero: "))
 
 if numero_entero >= 0:
@@ -37,8 +34,6 @@ else:
     print("Número negativo")
 
 
-# --- Bloque 4: Condición de borde (Rango 1 a 100) ---
-# Verificamos límites exactos y pertenencia al rango
 numero_rango = int(input("Ingresa un número entre 1 y 100: "))
 
 if numero_rango == 1 or numero_rango == 100:
